Adds iterable and (m)mappable data set objects
Training data is now loaded using an mmapped file, after making a single
pass through the file for offsets. In practice this should have little
effect on small files, but make very large files usable. There is a bit
of overhead, but this is mostly delegated to the (separate-process)
dataloader and ultimately to the kernel.

Other data files are now loaded iteratively, again delegating.